api/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
flask-cors==6.0.0
mysql-connector-python==9.2.0
python-dotenv==1.1.0
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
openai==1.81.0
pillow==11.2.1
requests==2.32.3
python-pptx==1.0.2
pytest==8.2.2
black==24.4.2
isort==5.12.0
